refactor: extract commom function to lib/helper.js
